website/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
google-generativeai==0.8.5
numpy==1.24.3
scikit-learn==1.3.2
pydantic==2.5.0